requirements.txt
Django==5.0.1
graphene-django==3.2.0
django-graphql-jwt==0.4.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
pytest==7.4.4
pytest-django==4.7.0